Move some tests from integration to internal
This is in preparation for making the Corrupt error type opaque. In
order to test specific errors, these tests must have access to crate
internals.
