+++ src/lib.rs
@@ -624,6 +624,55 @@ mod tests {
     use super::*;
     use test_util::load_test_disk1;
 
+    #[test]
+    fn test_load_errors() {
+        // Not enough data.
+        assert!(matches!(
+            Ext4::load(Box::new(vec![])).unwrap_err(),
+            Ext4Error::Io(_)
+        ));
+
+        // Invalid superblock.
+        assert!(matches!(
+            Ext4::load(Box::new(vec![0; 2048])).unwrap_err(),
+            Ext4Error::Corrupt(Corrupt::SuperblockMagic)
+        ));
+
+        // Not enough data to read the block group descriptors.
+        let mut fs_data = vec![0; 2048];
+        fs_data[1024..2048]
+            .copy_from_slice(include_bytes!("../test_data/raw_superblock.bin"));
+        assert!(matches!(
+            Ext4::load(Box::new(fs_data.clone())).unwrap_err(),
+            Ext4Error::Io(_)
+        ));
+
+        // Invalid block group descriptor checksum.
+        fs_data.resize(3048usize, 0u8);
+        assert!(matches!(
+            Ext4::load(Box::new(fs_data.clone())).unwrap_err(),
+            Ext4Error::Corrupt(Corrupt::BlockGroupDescriptorChecksum(0))
+        ));
+    }
+
+    /// Test that loading the data from
+    /// https://github.com/nicholasbishop/ext4-view-rs/issues/280 does not
+    /// panic.
+    #[test]
+    fn test_invalid_ext4_data() {
+        // Fill in zeros for the first 1024 bytes, then add the test data.
+        let mut data = vec![0; 1024];
+        data.extend(include_bytes!("../test_data/not_ext4.bin"));
+
+        assert_eq!(
+            *Ext4::load(Box::new(data))
+                .unwrap_err()
+                .as_corrupt()
+                .unwrap(),
+            Corrupt::InvalidBlockSize
+        );
+    }
+
     fn block_read_error(
         block_index: u64,
         offset_within_block: u32,
